+++ Giggity/db/migrate/20191204205823_create_venues.rb
@@ -4,7 +4,7 @@ class CreateVenues < ActiveRecord::Migration[6.0]
       t.string :username
       t.string :password_digest
       t.string :venue_name
-      t.integer :bankroll
+      t.float :bankroll
       t.integer :zipcode
       
       t.timestamps
